+++ packages/jobs/src/constants.ts
@@ -6,7 +6,6 @@ export const JOBS_PREFIX = "formbricks:jobs";
 export const JOB_NAMES = {
   testLog: "system.test-log",
   responsePipeline: "response-pipeline.process",
-  aiTranslation: "ai-translation.translate",
   surveyScheduling: "survey-scheduling.reconcile",
 } as const;
