+++ packages/jobs/src/definitions.ts
@@ -1,15 +1,9 @@
 import { JOB_NAMES } from "@/src/constants";
 import { type AnyBackgroundJobDefinition, toAnyBackgroundJobDefinition } from "@/src/contracts";
-import { processAITranslationJob } from "@/src/processors/ai-translation";
 import { processResponsePipelineJob } from "@/src/processors/response-pipeline";
 import { processSurveySchedulingJob } from "@/src/processors/survey-scheduling";
 import { processTestLogJob } from "@/src/processors/test-log";
-import {
-  ZAITranslationJobData,
-  ZResponsePipelineJobData,
-  ZSurveySchedulingJobData,
-  ZTestLogJobData,
-} from "@/src/types";
+import { ZResponsePipelineJobData, ZSurveySchedulingJobData, ZTestLogJobData } from "@/src/types";
 
 export const backgroundJobDefinitions = {
   [JOB_NAMES.responsePipeline]: toAnyBackgroundJobDefinition({
@@ -26,11 +20,6 @@ export const backgroundJobDefinitions = {
     handle: processTestLogJob,
     name: JOB_NAMES.testLog,
     schema: ZTestLogJobData,
-  }),
-  [JOB_NAMES.aiTranslation]: toAnyBackgroundJobDefinition({
-    handle: processAITranslationJob,
-    name: JOB_NAMES.aiTranslation,
-    schema: ZAITranslationJobData,
   }),
 } as const satisfies Record<string, AnyBackgroundJobDefinition>;
