+++ packages/jobs/src/types.ts
@@ -1,5 +1,4 @@
 import { z } from "zod";
-import { ZId } from "@formbricks/types/common";
 import { ZResponse } from "@formbricks/types/responses";
 import { ZTag } from "@formbricks/types/tags";
 
@@ -34,23 +33,6 @@ export const ZResponsePipelineJobData = z.object({
 });
 
 export type TResponsePipelineJobData = z.infer<typeof ZResponsePipelineJobData>;
-
-export const ZAITranslationField = z.object({
-  path: z.string(),
-  defaultText: z.string(),
-  isRichText: z.boolean(),
-});
-
-export const ZAITranslationJobData = z.object({
-  organizationId: z.string().min(1),
-  workspaceId: z.string().min(1),
-  userId: ZId,
-  fields: z.array(ZAITranslationField).min(1),
-  sourceLanguage: z.string().min(1),
-  targetLanguage: z.string().min(1),
-});
-
-export type TAITranslationJobData = z.infer<typeof ZAITranslationJobData>;
 
 export const ZSurveySchedulingJobData = z.object({
   scope: z.literal("global"),
